Update collections.mdx
Expand the description to explain the purpose of the Collections
Make the link to example more actionable and provide a description of what to expect (Example description).
Try to clarify description ( upsert for example).
Clarify "defaults to an empty namespace if not provided"
Change "to associate with the item" for get and search function: we don't associate a namespace when doing queries.
Misc clarification.